multiboot: moving headroom to per module_t
Hyperlaunch will require multiple modules to be allocated headroom. Therefore a
headroom variable has been added to module_t for tracking headroom on a per
module_t basis. To provide backwards compatibility with multiboot, the
reserved field was bitfield split to allow it to continue to be used to flag
module was relocated and to add to track the headroom needed when being
relocated.

RFC Note: The approach taken here was to produce something that was acceptable
but likely not the ultimate solution. A more complete solution would likely be
one that looked to unify a cross-architecture approach to represent the
material provided by bootloaders.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Reviewed-by: Christopher Clark <christopher.clark@starlab.io>